Перенесла Animal в стартер
