+++ src/main/resources/application.yml
@@ -1,0 +1,6 @@
+spring:
+  application.name: education-project
+
+name:
+  names:
+    animal: barsik, persik
