+++ starter/src/main/java/ru/mtsbank/NamesProperties.java
@@ -1,0 +1,10 @@
+package ru.mtsbank;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.List;
+
+@ConfigurationProperties("name")
+public class NamesProperties {
+    List<String> names;
+}
