+++ starter/src/main/java/ru/mtsbank/entity/Animal.java
@@ -1,4 +1,4 @@
-package ru.mts.hw_3.entity;
+package ru.mtsbank.entity;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
