+++ starter/src/main/java/ru/mtsbank/entity/Wolf.java
@@ -1,24 +1,25 @@
-package ru.mts.hw_3.entity;
+package ru.mtsbank.entity;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public class Dog extends Pet {
-    public Dog(String breed, String name, BigDecimal cost, String character, LocalDate birthDate) {
+public class Wolf extends Predator {
+    public Wolf(String breed, String name, BigDecimal cost, String character, LocalDate birthDate) {
         super(breed, name, cost, character, birthDate);
     }
 
     /**
-     * Метод - возвращает строку, представляющую объект Dog(Собака)
+     * Метод - возвращает строку, представляющую объект Wolf(Волк)
      */
     @Override
     public String toString() {
-        return "Dog {" +
+        return "Wolf {" +
                 "breed='" + breed + '\'' +
                 ", name='" + name + '\'' +
                 ", cost=" + cost +
                 ", character='" + character + '\'' +
-                ", birthDate=" + getBirthDateString() +
+                ", birthDate=" + getBirthDateString()
+                +
                 '}';
     }
 }
