+++ starter/src/main/java/ru/mtsbank/utility/Constants.java
@@ -1,0 +1,7 @@
+package ru.mtsbank.utility;
+
+import java.time.format.DateTimeFormatter;
+
+public class Constants {
+    public static final DateTimeFormatter DATE_FORMATTER_OUT = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+}
